Split event groups into separate files
